Factored out canvas resize and webgl setup to custom hooks
I previously got rid of a resize hook, but that was a different
approach than this one. This one actually works, and it can still
live in a hook to be a bit cleaner.
